refactor: changes jira-statuses to statuses + other platform support
