+++ src/commands/index.ts
@@ -2,7 +2,7 @@ export const commands = new Set([
     'files',
     'help',
     'issues',
-    'jira-statuses',
     'prepare',
     'projects',
+    'statuses',
   ]);
